+++ src/main/java/uk/gov/hmcts/reform/iacasenotificationsapi/domain/entities/MakeAnApplicationTypes.java
@@ -1,12 +1,16 @@
 package uk.gov.hmcts.reform.iacasenotificationsapi.domain.entities;
 
+import static java.util.Arrays.stream;
+
 import com.fasterxml.jackson.annotation.JsonValue;
+import java.util.Optional;
 
 public enum  MakeAnApplicationTypes {
 
     ADJOURN("Adjourn"),
     EXPEDITE("Expedite"),
     JUDGE_REVIEW("Judge's review of application decision"),
+    JUDGE_REVIEW_LO("Judge's review of Legal Officer decision"),
     LINK_OR_UNLINK("Link/unlink appeals"),
     TIME_EXTENSION("Time extension"),
     TRANSFER("Transfer"),
@@ -14,13 +18,26 @@ public enum  MakeAnApplicationTypes {
     UPDATE_HEARING_REQUIREMENTS("Update hearing requirements"),
     UPDATE_APPEAL_DETAILS("Update appeal details"),
     REINSTATE("Reinstate an ended appeal"),
+    TRANSFER_OUT_OF_ACCELERATED_DETAINED_APPEALS_PROCESS("Transfer out of accelerated detained appeals process"),
     OTHER("Other");
+
+    MakeAnApplicationTypes(String value) {
+        this.value = value;
+    }
 
     @JsonValue
     private final String value;
 
-    MakeAnApplicationTypes(String value) {
-        this.value = value;
+    public static Optional<MakeAnApplicationTypes> from(
+        String value
+    ) {
+        return stream(values())
+            .filter(v -> v.getValue().equals(value))
+            .findFirst();
+    }
+
+    public String getValue() {
+        return value;
     }
 
     @Override
